Separate barrier population into variants, share broadcast call
Worst commit ever
