+++ client/components/inline-barrier/main.js
@@ -1,130 +1,26 @@
-/**
-* !WIP Proof of Concept
-*
-* For now we simply render an inline barrier page similar to fastft.
-* In future this could become an n/o-component
-*/
-
-import { broadcast } from 'n-ui/utils';
-
-const broadcastOpportunity = (subtype) => {
-	const appNameEl = document.querySelector('[data-next-app]');
-	const appVersionEl = document.querySelector('[data-next-version]');
-	const context = {
-		product: 'next',
-		app: appNameEl && appNameEl.getAttribute('data-next-app'),
-		appVersion: appVersionEl && appVersionEl.getAttribute('data-next-version')
-	};
-	const acquisitionContextEls = document.querySelectorAll('[data-acquisition-context]');
-	const offersEls = document.querySelectorAll('[data-offer-id]');
-
-	broadcast('oTracking.event', Object.assign({
-		category: 'barrier',
-		action: 'view',
-		opportunity: {
-			type: 'barrier',
-			subtype
-		},
-		type: 'barrier',
-		acquisitionContext: Array.from(acquisitionContextEls).map(e => e.getAttribute('data-acquisition-context')),
-		offers: Array.from(offersEls).map(e => e.getAttribute('data-offer-id'))
-	}, context))
-};
-
-const populateWithPsp = (el) => fetch('/products?fragment=true&inline=true&narrow=true&in-article=true', { credentials: 'same-origin' })
-	.then(response => {
-		if (response.ok) {
-			return response.text().then(html => {
-				const parsedHtml = new DOMParser().parseFromString(html, 'text/html');
-				const coreWrapper = el.querySelector('.inline-barrier__content');
-				const pspEl = parsedHtml.querySelector('.barrier');
-
-				if (!el || !pspEl || !coreWrapper) { return; }
-
-				//Remove duplicate ID so pa11y passes
-				pspEl.removeAttribute('id');
-				const pspTitle = pspEl.querySelector('.barrier-grid__heading-pretext');
-
-				if (pspTitle) {
-					// Horrible hacks to modify the PSP title text, for a slightly dirty MVT
-					const titleContainer = pspTitle.parentElement;
-					const nonH1Title = document.createElement('p');
-					nonH1Title.className = 'barrier-grid__heading barrier-grid__heading-pretext';
-					nonH1Title.innerHTML = 'Want to read more?';
-					const sub = document.createElement('p');
-					sub.innerHTML = '<span class="barrier-grid__article-title">Subscribe by choosing a package below:</span>';
-					titleContainer.insertAdjacentElement('beforeend', nonH1Title);
-					titleContainer.insertAdjacentElement('beforeend', sub);
-					pspTitle.remove();
-				}
-
-				el.classList.remove('inline-barrier--no-prices');
-				el.classList.add('inline-barrier--has-prices');
-				el.replaceChild(pspEl, coreWrapper);
-
-				broadcastOpportunity('inline-psp');
-			});
-		}
-	})
-	.catch(() => {
-		// eslint-disable-next-line no-console
-		console.error('Failed to retrieve/process PSP fragment');
-	})
-	.then(() => {
-		el.classList.add('inline-barrier--done');
-	});
-
-const populateWithTrial = (el) => fetch(`https://next-signup-api.ft.com/offer/41218b9e-c8ae-c934-43ad-71b13fcb4465?countryCode=${el.dataset.countryCode}`, {
-	headers: {
-		'x-api-env': 'prod'
-	}
-})
-	.then(response => {
-		if (response.ok) {
-			return response.json();
-		} else {
-			throw new Error('Network response was not ok.');
-		}
-	})
-	.then(json => json.data)
-	.then(({offer}) => {
-		const coreWrapper = el.querySelector('.inline-barrier__content');
-		const enhancedTrialTemplate = document.querySelector('#js-barrier-trial');
-
-		if (!el || !enhancedTrialTemplate || !coreWrapper) { return; }
-
-		coreWrapper.innerHTML = enhancedTrialTemplate.innerHTML;
-		const trialPriceables = [...coreWrapper.querySelectorAll('.js-barrier-trial-price')];
-		const postTrialPriceables = [...coreWrapper.querySelectorAll('.js-barrier-trial-after-price')];
-
-		const trialOffer = offer.charges.find(charge => charge.billing_period === 'trial');
-		//FIXME: make sure this is the right object: why different amounts from psp page?
-		const postTrialOffer = offer.charges.find(charge => charge.billing_period === 'monthly');
-		//TODO: investigate localization of symbol/value ordering
-
-		trialPriceables.forEach(priceable => priceable.innerHTML = `${trialOffer.amount.symbol}${trialOffer.amount.value.replace('.00', '')}`);
-		postTrialPriceables.forEach(priceable => priceable.innerHTML = `${postTrialOffer.display_amount.weekly.symbol}${postTrialOffer.display_amount.weekly.value}`);
-
-		el.classList.remove('inline-barrier--no-prices');
-		el.classList.add('inline-barrier--has-prices');
-
-		broadcastOpportunity('inline-trial');
-	})
-	.catch(() => {
-		// eslint-disable-next-line no-console
-		console.error('Failed to retrieve/process trial offer data');
-	})
-	.then(() => {
-		el.classList.add('inline-barrier--done');
-	});
+import populateWithPsp from './populate-with-psp';
+import populateWithTrial from './populate-with-trial';
+import broadcastOpportunity from './broadcast-opportunity';
 
 export default (flags) => {
 	const el = document.querySelector('.inline-barrier');
+	const inArticlePreviewVariant = flags.get('inArticlePreview');
 
-	if (el && el.dataset.nType === 'standard' && flags.get('inArticlePreview') === 'psp') {
-		populateWithPsp(el);
-	}
-	else if (el && el.dataset.nType === 'standard' && el.dataset.countryCode && flags.get('inArticlePreview') === 'trial') {
-		populateWithTrial(el);
-	}
+	if (!el || !inArticlePreviewVariant || !el.dataset.countryCode) { return; }
+
+	const populateWithOffer = (inArticlePreviewVariant === 'trial') ? populateWithTrial : populateWithPsp;
+
+	populateWithOffer(el)
+		.then(() => {
+			// Tracking for MVT analysis
+			broadcastOpportunity(`inline-${inArticlePreviewVariant}`);
+		})
+		.catch((err) => {
+			// eslint-disable-next-line no-console
+			console.error('Failed to retrieve/process offer data', err);
+		})
+		.then(() => {
+			// Show either the core or enhanced barrier, depending on success of fetching offer data
+			el.classList.add('inline-barrier--done');
+		});
 }
