+++ views/partials/inline-barrier/template.html
@@ -6,7 +6,7 @@
 		<img src="https://www.ft.com/__origami/service/image/v2/images/raw/https%3A%2F%2Fwww.ft.com%2F__assets%2Fcreatives%2Fstandard_digital.png?width=160&amp;source=next&amp;fit=scale-down" alt="" role="presentation" aria-hidden="true">
 	{{/inline}}
 
-	<div class="inline-barrier inline-barrier--{{@root.flags.inArticlePreview}} inline-barrier--no-prices" data-n-component="inline-barrier" data-n-type="{{type}}" data-country-code="{{countryCode}}" data-trackable="inline-barrier">
+	<div class="inline-barrier inline-barrier--{{@root.flags.inArticlePreview}}" data-n-component="inline-barrier" data-n-type="{{type}}" data-country-code="{{countryCode}}" data-trackable="inline-barrier">
 		<div class="inline-barrier__fader">
 			<p class="n-util-visually-hidden">This is the end of your free article preview.</p>
 		</div>
